+++ mensasverige/features/events/utils/eventUtils.ts
@@ -1,20 +1,5 @@
 import { Event } from '../../../api_schema/types';
-
-export interface EventFilter {
-  attendingOrHost?: boolean;
-  upcoming?: boolean;
-  limit?: number;
-}
-
-/**
- * Extended Event type with computed properties
- */
-export interface ExtendedEvent extends Event {
-  /** Whether the user is attending or is a host/admin */
-  attendingOrHost: boolean;
-  /** Whether the event is in the future */
-  isFutureEvent: boolean;
-}
+import { EventFilter, ExtendedEvent, GroupedEvents } from '../types/eventUtilTypes';
 
 /**
  * Check if user is effectively attending an event (including admin/host status)
@@ -57,34 +42,12 @@ export const createExtendedEvent = (event: Event, currentUserId?: number): Exten
   };
 };
 
-/**
- * Convert an array of Events to ExtendedEvents with computed properties
- */
-export const createExtendedEvents = (events: Event[], currentUserId?: number): ExtendedEvent[] => {
-  return events.map(event => createExtendedEvent(event, currentUserId));
-};
-
-export interface GroupedEvents {
-  [key: string]: ExtendedEvent[];
-}
-
 export const displayLocaleTimeStringDate = (datestring: string) => {
     const date: Date = new Date(datestring ?? "");
     const weekday = date.toLocaleDateString('sv-SE', { weekday: 'long' });
     const capitalizedWeekday = weekday.charAt(0).toUpperCase() + weekday.slice(1);
     const dayMonth = date.toLocaleDateString('sv-SE', { day: 'numeric', month: 'long' });
     return `${capitalizedWeekday} ${dayMonth}`;
-};
-
-export const createUpcomingEventsFilter = () => {
-    return (event: Event) => {
-        if (!event.start || !event.end || !event.attending) return false;
-        const now = new Date();
-        const eventDate = new Date(event.start);
-        const [hours, minutes] = event.end.split(':');
-        eventDate.setHours(parseInt(hours), parseInt(minutes));
-        return eventDate > now;
-    };
 };
 
 /**
@@ -157,71 +120,4 @@ export const groupEventsByDate = (events: ExtendedEvent[]): GroupedEvents => {
   }, {} as GroupedEvents);
 };
 
-/**
- * Process events with filtering, sorting, and grouping
- */
-export const processEvents = (events: ExtendedEvent[], filter: EventFilter = {}): GroupedEvents => {
-  const filteredEvents = filterEvents(events, filter);
-  return groupEventsByDate(filteredEvents);
-};
 
-/**
- * Find the next upcoming event from grouped events
- */
-export const findNextEvent = (groupedEvents: GroupedEvents, excludeFirstEvent: boolean = false): ExtendedEvent | null => {
-  if (!groupedEvents || Object.keys(groupedEvents).length === 0) return null;
-
-  const now = new Date();
-  let nextEventTemp: ExtendedEvent | null = null;
-
-  for (const date in groupedEvents) {
-    for (const event of groupedEvents[date]) {
-      if (!event.start) continue;
-
-      const eventDate = new Date(event.start);
-      if (eventDate > now) {
-        nextEventTemp = event;
-        break;
-      }
-    }
-    if (nextEventTemp) break;
-  }
-
-  // Optionally exclude the first event if it's the next event
-  if (excludeFirstEvent && nextEventTemp) {
-    const firstEventOfFirstDay = groupedEvents[Object.keys(groupedEvents)[0]][0];
-    if (nextEventTemp.id === firstEventOfFirstDay.id) {
-      return null;
-    }
-  }
-
-  return nextEventTemp;
-};
-
-/**
- * Get upcoming events with attending filter for dashboard
- */
-export const getUpcomingAttendingEvents = (events: ExtendedEvent[], limit: number = 3): {
-  groupedEvents: GroupedEvents;
-  hasMore: boolean;
-} => {
-  const upcomingAttendingEvents = filterEvents(events, { 
-    attendingOrHost: true, 
-    upcoming: true 
-  });
-  
-  const limitedEvents = upcomingAttendingEvents.slice(0, limit);
-  const groupedEvents = groupEventsByDate(limitedEvents);
-  
-  return {
-    groupedEvents,
-    hasMore: upcomingAttendingEvents.length > limit
-  };
-};
-
-/**
- * Get all events processed and grouped (for schedule view)
- */
-export const getAllEventsGrouped = (events: ExtendedEvent[]): GroupedEvents => {
-  return processEvents(events);
-};
